Support for different uploaders
