+++ .env.example
@@ -36,12 +36,29 @@ TENOR=
 
 # Put HTML help page output location here, leave blank to disable
 OUTPUT=
-# Put temporary image dir here (make sure it's accessible via a web server), leave blank to disable
-TEMPDIR=
-# Put temporary image web server domain
+
+# Where to upload files too large to send directly to Discord
+# If set to "tempdir", files will be written to $TEMPDIR variable.
+# If set to "s3", files will be uploaded to an S3-compatible server.
+#
+# If it is unset but $TEMPDIR is set, this will default to "tempdir" for
+# backwards compatibility reasons.
+UPLOADER=
+# For tempdir and s3 uploaders, web server domain for image hosting
 TMP_DOMAIN=
 # Threshold where optional space saving methods will be performed
 THRESHOLD=
+
+# For tempdir uploader, put temporary image dir here (make sure it's accessible via a web server)
+TEMPDIR=
+
+# Options for s3 uploader
+S3_ENDPOINT=
+S3_ACCESS_KEY=
+S3_SECRET_KEY=
+S3_BUCKET=
+S3_NO_SSL=
+
 # Port for serving metrics. Metrics served are compatible with Prometheus.
 METRICS=
