+++ events/interactionCreate.js
@@ -1,5 +1,6 @@
 import { promises } from "fs";
 import database from "../utils/database.js";
+import uploader from "../utils/uploader.js"
 import * as logger from "../utils/logger.js";
 import { commands } from "../utils/collections.js";
 import { CommandInteraction } from "eris";
@@ -38,10 +39,8 @@ export default async (client, cluster, worker, ipc, interaction) => {
         }
       }
       if (result.file.length > fileSize) {
-        if (process.env.TEMPDIR && process.env.TEMPDIR !== "") {
-          const filename = `${Math.random().toString(36).substring(2, 15)}.${result.name.split(".")[1]}`;
-          await promises.writeFile(`${process.env.TEMPDIR}/${filename}`, result.file);
-          const imageURL = `${process.env.TMP_DOMAIN || "https://tmp.projectlounge.pw"}/${filename}`;
+        if (uploader) {
+          const imageURL = await uploader.upload(message.id, result);
           await interaction[interaction.acknowledged ? "editOriginalMessage" : "createMessage"]({
             embeds: [{
               color: 16711680,
@@ -55,30 +54,6 @@ export default async (client, cluster, worker, ipc, interaction) => {
               },
             }]
           });
-          if (process.env.THRESHOLD) {
-            process.env.DIRSIZECACHE += result.file.length;
-            if (process.env.DIRSIZECACHE > process.env.THRESHOLD) {
-              const files = (await promises.readdir(process.env.TEMPDIR)).map((file) => {
-                return new Promise((resolve, reject) => {
-                  promises.stat(`${process.env.TEMPDIR}/${file}`).then((fstats)=>{
-                    resolve({
-                      name: file,
-                      size: fstats.size,
-                      ctime: fstats.ctime
-                    });
-                  }).catch(reject);
-                });
-              });
-              Promise.all(files).then((files) => {
-                process.env.DIRSIZECACHE = files.reduce((a, b)=>{
-                  return a+b.size;
-                }, 0);
-                const oldestFile = files.sort((a, b) => a.ctime - b.ctime)[0].name;
-                promises.rm(`${process.env.TEMPDIR}/${oldestFile}`);
-                logger.log(`Removed oldest image file: ${oldestFile}`);
-              });
-            }
-          }
         } else {
           await interaction[interaction.acknowledged ? "editOriginalMessage" : "createMessage"]("The resulting image was more than 8MB in size, so I can't upload it.");
         }
