+++ events/messageCreate.js
@@ -1,10 +1,10 @@
 import { promises } from "fs";
 import database from "../utils/database.js";
+import uploader from "../utils/uploader.js";
 import { log, error as _error } from "../utils/logger.js";
 import { prefixCache, aliases, disabledCache, disabledCmdCache, commands } from "../utils/collections.js";
 import parseCommand from "../utils/parseCommand.js";
 import { clean } from "../utils/misc.js";
-
 // run when someone sends a message
 export default async (client, cluster, worker, ipc, message) => {
   // ignore other bots
@@ -125,10 +125,8 @@ export default async (client, cluster, worker, ipc, message) => {
         }
       }
       if (result.file.length > fileSize) {
-        if (process.env.TEMPDIR && process.env.TEMPDIR !== "") {
-          const filename = `${Math.random().toString(36).substring(2, 15)}.${result.name.split(".")[1]}`;
-          await promises.writeFile(`${process.env.TEMPDIR}/${filename}`, result.file);
-          const imageURL = `${process.env.TMP_DOMAIN || "https://tmp.projectlounge.pw"}/${filename}`;
+        if (uploader) {
+          const imageURL = await uploader.upload(message.id, result);
           await client.createMessage(message.channel.id, Object.assign({
             embeds: [{
               color: 16711680,
@@ -142,30 +140,6 @@ export default async (client, cluster, worker, ipc, message) => {
               },
             }]
           }, reference));
-          if (process.env.THRESHOLD) {
-            process.env.DIRSIZECACHE += result.file.length;
-            if (process.env.DIRSIZECACHE > process.env.THRESHOLD) {
-              const files = (await promises.readdir(process.env.TEMPDIR)).map((file) => {
-                return new Promise((resolve, reject) => {
-                  promises.stat(`${process.env.TEMPDIR}/${file}`).then((fstats)=>{
-                    resolve({
-                      name: file,
-                      size: fstats.size,
-                      ctime: fstats.ctime
-                    });
-                  }).catch(reject);
-                });
-              });
-              Promise.all(files).then((files) => {
-                process.env.DIRSIZECACHE = files.reduce((a, b)=>{
-                  return a+b.size;
-                }, 0);
-                const oldestFile = files.sort((a, b) => a.ctime - b.ctime)[0].name;
-                promises.rm(`${process.env.TEMPDIR}/${oldestFile}`);
-                log(`Removed oldest image file: ${oldestFile}`);
-              });
-            }
-          }
         } else {
           await client.createMessage(message.channel.id, "The resulting image was more than 8MB in size, so I can't upload it.");
         }
