+++ package.json
@@ -32,6 +32,7 @@
     "file-type": "^17.1.2",
     "format-duration": "^2.0.0",
     "jsqr": "^1.4.0",
+    "minio": "^7.0.28",
     "node-addon-api": "^5.0.0",
     "node-emoji": "^1.11.0",
     "node-fetch": "^3.2.6",
